feat: add theme switcher config [MDS-1644]
