+++ docs/.storybook/preview-body.html
@@ -1,7 +1,5 @@
 <html dir="ltr">
-  <body
-    class="light-theme theme-moon-light flex items-center justify-center h-screen"
-  >
+  <body class="flex items-center justify-center h-screen">
     <div id="storybook-root"></div>
     <div id="headlessui-portal-root"><div></div></div>
   </body>
